+++ src/base/window_manager.c
@@ -305,7 +305,8 @@ widget_t* window_manager_find_target(widget_t* widget, xy_t x, xy_t y) {
   xy_t r = iter->x + iter->w;
   xy_t b = iter->y + iter->h;
 
-  if (iter->visible && iter->enable && p.x >= iter->x && p.y >= iter->y && p.x <= r && p.y <= b) {
+  if (iter->visible && iter->sensitive && iter->enable && p.x >= iter->x && p.y >= iter->y &&
+      p.x <= r && p.y <= b) {
     return iter;
   }