+++ tests/button_test.cc
@@ -39,6 +39,8 @@ TEST(Button, clone) {
   ASSERT_EQ(widget_set_prop(w1, WIDGET_PROP_REPEAT, &v1), RET_OK);
   widget_set_self_layout_params(w1, "1", "2", "3", "4");
   widget_set_children_layout(w1, "default(r=0,c=0,x=10,y=10,s=10)");
+  widget_set_sensitive(w1, FALSE);
+  widget_set_floating(w1, TRUE);
   ASSERT_EQ(button_cast(w1), w1);
 
   widget_t* w2 = widget_clone(w1, NULL);