+++ tests/ui_serializer_test.cc
@@ -33,8 +33,8 @@ TEST(UISerializer, layout_self) {
   widget_to_xml(w, &str);
 
   ASSERT_EQ(string(str.str),
-            "<button x=\"10\" y=\"20\" w=\"30\" h=\"40\" repeat=\"0\" "
-            "self_layout=\"default(x=r:100,y=m:10,w=0,h=10%)\">\n</button>\n");
+            "<button x=\"10\" y=\"20\" w=\"30\" h=\"40\" "
+            "self_layout=\"default(x=r:100,y=m:10,w=0,h=10%)\" repeat=\"0\">\n</button>\n");
 
   str_reset(&str);
   widget_destroy(w);
@@ -49,8 +49,8 @@ TEST(UISerializer, layout_self1) {
   widget_to_xml(w, &str);
 
   ASSERT_EQ(string(str.str),
-            "<button x=\"10\" y=\"20\" w=\"30\" h=\"40\" repeat=\"0\" "
-            "self_layout=\"default(x=c:100,y=b:10,w=0,h=10%)\">\n</button>\n");
+            "<button x=\"10\" y=\"20\" w=\"30\" h=\"40\" "
+            "self_layout=\"default(x=c:100,y=b:10,w=0,h=10%)\" repeat=\"0\">\n</button>\n");
 
   str_reset(&str);
   widget_destroy(w);