+++ docs/changes.md
@@ -6,6 +6,8 @@
   * 子控件支持浮动布局，不受父控件的children\_layouter的限制。
   * 支持等宽字体的问题(感谢[zjm1060](https://github.com/zjm1060)提供补丁)。 
   * 修改vgcanvas中w/h为负数时的BUG(感谢朝泽提供补丁)。
+  * 完善widget\_clone/widget\_equal/widget\_to\_xml。
+  * widget增加sensitive属性。
 
 * 2018/12/19
   * 修改combobox弹出窗口位置的BUG。